cloud/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncmy==0.2.9
pydantic==2.10.3
pydantic-settings==2.7.1
python-ulid==3.0.0
python-dotenv==1.0.1
structlog==24.4.0
websockets==14.1
httpx==0.28.1
